Split paper/etc constraint migration from others

+++ db/migrate/20140304135944_remove_old_tables.rb
@@ -1,10 +1,14 @@
 class RemoveOldTables < ActiveRecord::Migration
-  def change
-    drop_table :delayed_jobs
-    drop_table :down_votes
-    drop_table :downvotes
-    drop_table :feed_days
-    drop_table :up_votes
-    drop_table :upvotes
+  def up
+    drop_table :delayed_jobs if table_exists? :delayed_jobs
+    drop_table :down_votes if table_exists? :down_votes
+    drop_table :downvotes if table_exists? :downvotes
+    drop_table :feed_days if table_exists? :feed_days
+    drop_table :up_votes if table_exists? :up_votes
+    drop_table :upvotes if table_exists? :upvotes
+  end
+
+  def down
+    raise
   end
 end
